+++ source/components/cardContainer/card/headerColumn/headerColumn.js
@@ -21,7 +21,7 @@ function headerColumn($compile) {
     'use strict';
     return {
         restrict: 'E',
-        template: "\n\t\t\t<div size-for-breakpoints=\"header.column.size\" title=\"{{::header.column.description}}\">\n\t\t\t\t<div class=\"template-container\" style=\"display: inline-block\"></div>\n\t\t\t</div>\n\t\t",
+        template: "\n\t\t\t<div rl-size-for-breakpoints=\"header.column.size\" title=\"{{::header.column.description}}\">\n\t\t\t\t<div class=\"template-container\" style=\"display: inline-block\"></div>\n\t\t\t</div>\n\t\t",
         controller: exports.controllerName,
         controllerAs: 'header',
         scope: {},
